+++ test/cases/option-preload-attributes/webpack.config.js
@@ -1,0 +1,56 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  resolve: {
+    alias: {
+      '@fonts': path.join(__dirname, 'src/assets/fonts'),
+    },
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: './src/views/index.html',
+      },
+
+      css: {
+        filename: 'assets/css/[name].[contenthash:8].css',
+      },
+
+      preload: [
+        {
+          test: /\.(s?css|less)$/,
+          as: 'style',
+        },
+        {
+          test: /\.(eot|ttf|woff2?)$/,
+          attributes: { as: 'font', crossorigin: true }, // test `as` in attributes
+        },
+      ],
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.(css|sass|scss)$/,
+        use: ['css-loader', 'sass-loader'],
+      },
+
+      {
+        test: /\.(eot|ttf|woff2?)$/,
+        type: 'asset/resource',
+        generator: {
+          filename: 'assets/fonts/[name][ext]',
+        },
+      },
+    ],
+  },
+};