+++ test/cases/split-chunk-resolve-assets-dev/expected/assets/js/scripts-src_assets_js_main-a_js.e4a86eee.js
@@ -6,7 +6,7 @@
  * or disable the default devtool with "devtool: false".
  * If you are looking for production-ready output files, see mode: "production" (https://webpack.js.org/configuration/mode/).
  */
-(self["webpackChunk"] = self["webpackChunk"] || []).push([["scripts-src_assets_js_main-a_js-_8b421"],{
+(self["webpackChunk"] = self["webpackChunk"] || []).push([["scripts-src_assets_js_main-a_js"],{
 
 /***/ "./src/assets/js/main-a.js":
 /*!*********************************!*\
@@ -15,6 +15,16 @@
 /***/ ((__unused_webpack_module, __unused_webpack_exports, __webpack_require__) => {
 
 eval("const modA = __webpack_require__(/*! ./module-a */ \"./src/assets/js/module-a.js\");\nconst modC = __webpack_require__(/*! ./module-c */ \"./src/assets/js/module-c.js\");\n\nconsole.log('>> main-a:');\nconsole.log(' - A: ', modA);\nconsole.log(' - C: ', modC);\n\n//# sourceURL=webpack:///./src/assets/js/main-a.js?");
+
+/***/ }),
+
+/***/ "./src/assets/js/module-a.js":
+/*!***********************************!*\
+  !*** ./src/assets/js/module-a.js ***!
+  \***********************************/
+/***/ ((module, __unused_webpack_exports, __webpack_require__) => {
+
+eval("const lib = __webpack_require__(/*! ./lib */ \"./src/assets/js/lib.js\");\nconst value = lib.methodA();\n\nmodule.exports = value;\n\n//# sourceURL=webpack:///./src/assets/js/module-a.js?");
 
 /***/ })