+++ test/webpack.common.js
@@ -7,11 +7,4 @@ module.exports = {
   module: {
     rules: [],
   },
-
-  optimization: {
-    removeEmptyChunks: true,
-    mergeDuplicateChunks: true,
-    usedExports: true,
-    concatenateModules: true,
-  },
 };